+++ NEWS.md
@@ -9,6 +9,9 @@
 * `RETRY()` now throws the correct error message if an error occurs during the 
   request (@austin3dickey, #581).
 
+* New `download_file()` function to support cache-aware single and multi-URL
+  file downloads. (@hrbrmstr)
+
 # httr 1.4.0
 
 ## OAuth